+++ pkg/gatewayserver/io/mqtt/mqtt_test.go
@@ -301,6 +301,7 @@ func TestTraffic(t *testing.T) {
 						UplinkToken: io.MustUplinkToken(
 							ttnpb.GatewayAntennaIdentifiers{GatewayIdentifiers: registeredGatewayID},
 							100,
+							100000,
 							time.Unix(0, 100*1000),
 						),
 					},
@@ -328,6 +329,7 @@ func TestTraffic(t *testing.T) {
 						UplinkToken: io.MustUplinkToken(
 							ttnpb.GatewayAntennaIdentifiers{GatewayIdentifiers: registeredGatewayID},
 							100,
+							100000,
 							time.Unix(0, 100*1000),
 						),
 					},
@@ -358,6 +360,7 @@ func TestTraffic(t *testing.T) {
 						UplinkToken: io.MustUplinkToken(
 							ttnpb.GatewayAntennaIdentifiers{GatewayIdentifiers: registeredGatewayID},
 							100,
+							100000,
 							time.Unix(0, 100*1000),
 						),
 					},